backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic
python-dotenv
